// src/config/types/network-types.ts
export interface INetworkUrls {
    ETHEREUM_MAINNET: string;
    ETHEREUM_SEPOLIA: string;
    ETHEREUM_GOERLI: string;
    OPTIMISM_MAINNET: string;
    OPTIMISM_GOERLI: string;
    BASE_MAINNET: string;
    BASE_GOERLI: string;
    ZORA_MAINNET: string;
    ZORA_GOERLI: string;
    LOCALHOST: string;
}